requirements.txt
Pillow==8.3.1
imageio==2.9.0
opencv-python==4.5.3.56
PyQt5==5.15.4
python-tr==0.1.2
ffmpeg-python==0.2.0
moviepy==1.0.3